feat(cli): add secure-server command for firewall and SSH hardening
- Introduced `secure-server` CLI command to automate basic server security
- Enables UFW and allows specified ports via --allow-ports
- Disables SSH root login with --disable-root-login
- Enforces SSH key-based auth with --enforce-ssh-key-only
- Restarts SSH service after changes
- Skips existing UFW rules cleanly
