+++ cmd/ssl/check_expiry.go
@@ -1,34 +1,39 @@
 package ssl
 
 import (
+	"crypto/tls"
 	"fmt"
 	"os"
-	"crypto/tls"
 	"time"
+
 	"github.com/spf13/cobra"
 	"stackroost/internal/logger"
 )
 
-var CheckSSLExpiryCmd = &cobra.Command{
-	Use:   "check-ssl-expiry",
-	Short: "Check the SSL certificate expiry for a domain",
-	Run: func(cmd *cobra.Command, args []string) {
-		domain, _ := cmd.Flags().GetString("domain")
-		if domain == "" {
-			logger.Error("Please provide a domain using --domain")
-			os.Exit(1)
-		}
-		checkSSLExpiry(domain)
-	},
-}
+// GetExpiryCmd returns the cobra command to check SSL expiry
+func GetExpiryCmd() *cobra.Command {
+	cmd := &cobra.Command{
+		Use:   "expiry",
+		Short: "Check the SSL certificate expiry for a domain",
+		Run: func(cmd *cobra.Command, args []string) {
+			domain, _ := cmd.Flags().GetString("domain")
+			if domain == "" {
+				logger.Error("Please provide a domain using --domain")
+				os.Exit(1)
+			}
+			checkSSLExpiry(domain)
+		},
+	}
 
-func init() {
-	CheckSSLExpiryCmd.Flags().String("domain", "", "Domain to check SSL expiry for")
-	CheckSSLExpiryCmd.MarkFlagRequired("domain")
+	cmd.Flags().String("domain", "", "Domain to check SSL expiry for")
+	cmd.MarkFlagRequired("domain")
+	return cmd
 }
 
 func checkSSLExpiry(domain string) {
-	conn, err := tls.Dial("tcp", domain+":443", nil)
+	conn, err := tls.Dial("tcp", domain+":443", &tls.Config{
+		ServerName: domain, // Required for SNI
+	})
 	if err != nil {
 		logger.Error(fmt.Sprintf("Failed to connect: %v", err))
 		os.Exit(1)
